+++ src/dictionary/DefaultEdit.ts
@@ -29,7 +29,7 @@ export interface DefaultEditInitOption extends DefaultModInitOption {
 
 class DefaultEdit extends DefaultMod {
   static $name = 'DefaultEdit'
-  static $formatConfig = { name: 'Data:DefaultEdit', level: 50, recommend: true }
+  static $formatConfig = { name: 'DefaultEdit', level: 50, recommend: true }
   static $defaultValue = function(multiple: boolean) {
     return !multiple ? undefined : []
   }